pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bike_energy"
version = "1.0.0"
description = "Simulate cyclist energy consumption on road segments"
authors = [{ name = "Alexander Gustafsson & Eddie Tunas Ericson" }]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "gpxpy",
  "matplotlib",
  "numpy",
  "pyyaml",
  "pyproj",
  "scipy",
]

[tool.setuptools.packages.find]
where = ["src"]
